add VirtualList to Dropdown to improve perf of long lists
